parens: add support for keeping extraneous parens
Mercurial uses extraneous parens as a low-cost way to mark strings as
"intentionally not translated with _()". It's inelegant, but adding
this command-line flag allowed me to run pyupgrade on Mercurial,
fixing a bunch of set/dict comprehensions, set literals, and even
fixing a bug in a regex that was missing an r-prefix on the string.