Add `--check` option
This adds a `pyupgrade --check` option that does only check but not
rewrite the given files. If a file would be rewritten, a message is
printed to `stderr`.

Alternatives
------------

1. Using [pre-commit](https://pre-commit.com/)

Instead of adding this option, the user could use
`pre-commit run --all-files --show-diff-on-failure`.

2. Querying the version control system

Instead of adding this option, the user could let `pyupgrade` change the
files and use `git` or a similar tool to determine if something has
changed. After that the version control system could revert the changes.

3. Writing an orchestration script

Instead of adding this option, the user could write their own
orchestration script that would open and read the file to be checked and
pass them via `stdin` to `pyupgrade`. It could then compare the output
to the original content.

Why this is useful
------------------

A `--check` option is useful whenever one wants to enforce that all code
is compliant to a certain standard. This applies to `git` `pre-commit`
hooks, but also automated tests (especially run in a CI context).

Adding this option provides a significant benefit while adding minimal
additional maintenaince burden. The alternatives mentioned above all
have their own shortcomings:

1. In certain restricted environments it is not possible to load
arbitrary code during a CI run. Especially not from untrusted sources
(`pypi`, `npm` or even github). Additionally in some environments the
user might want to lock down exact versions of tools to be used. While
possible with `pre-commit`, it requires quite some overhead to install
and configure `pre-commit` and create and maintain a local
hook-repository just to be able to run `pyupgrade` as part of an
automated CI test.

2. This is simply not a good idea when it comes to parallel test
execution when we imagine that multiple tools modify files.

3. This is an actual alternative with two problems: It requires every
user to write this orchestration themselves while providing a `--check`
option is minimally intrusive. It requires a `pyupgrade` invocation for
every file to be checked. This can become quite costly in environments
with many python files.

Alternative implementation
--------------------------

Instead of just printing the filename that would be rewritten, the tool
could also print a diff that informs the user about the changes to be
made. This idea was discarded as it introduces a higher maintenaince
burden.

```python
import difflib

print(
    '\n'.join(
        difflib.unified_diff(
            contents_text_orig.splitlines(),
            contents_text.splitlines(),
            fromfile=f'a/{filename}',
            tofile=f'b/{filename}',
        )
    )
)
```

Related
-------

- Add --check option #210
- Feature request: dry-run option #258
- Add --check flag #310
- Provide better options to integrate with CI #356